requirements.txt
streamlit
langchain
langchain-community
langchain-core
langsmith
python-dotenv
ollama